refactor(ui): move dialog components into sync-dialogs/components/, slim tab to public API
Finish the sync-dialogs.tsx decomposition. The three subcomponents move
into sync-dialogs/components/: SyncDialogHost (RadixDialog shell +
auto-focus + open-change resolver wiring), SyncDialogBody (shared body
for all three kinds with input-field validation), and
DuplicatePersonDialogContent (two-step same-person vs merge flow).
sync-dialogs.tsx now owns only the dialogMount singleton and the three
exported entrypoints (ensureSyncDialogHost, openSyncConfirmDialog,
openSyncInputDialog, openDuplicatePersonDialog) plus the
DuplicatePersonActorOption type re-export consumed by team-sync.
