+++ packages/ui/src/tabs/sync/sync-dialogs.tsx
@@ -1,18 +1,14 @@
+/* Sync dialog queue — public entrypoints for the Sync and
+ * coordinator-admin flows to open a single in-flight confirm, input,
+ * or duplicate-person dialog. Implementation lives under
+ * sync-dialogs/ (types, internal state, and the three render
+ * components). */
+
 import { render } from "preact";
-import { useEffect, useMemo, useState } from "preact/hooks";
-import { DialogCloseButton } from "../../components/primitives/dialog-close-button";
-import { RadixDialog } from "../../components/primitives/radix-dialog";
-import { RadixRadioGroup } from "../../components/primitives/radix-radio-group";
-import { RadixSelect } from "../../components/primitives/radix-select";
-import { TextInput } from "../../components/primitives/text-input";
+import { SyncDialogHost } from "./sync-dialogs/components/sync-dialog-host";
 import {
-	dialogToneClassName,
 	ensureDialogAvailable,
 	fallbackResult,
-	getCurrentRequest,
-	isHostRequestSetter,
-	resolveCurrentDialog,
-	setHostRequestSetter,
 	setRequest,
 	setResolveDialog,
 } from "./sync-dialogs/internal";
@@ -21,336 +17,11 @@ import type {
 	DuplicatePersonDialogRequest,
 	DuplicatePersonDialogResult,
 	InputDialogRequest,
-	SyncDialogRequest,
 } from "./sync-dialogs/types";
 
 export type { DuplicatePersonActorOption } from "./sync-dialogs/types";
 
 let dialogMount: HTMLElement | null = null;
-
-function SyncDialogHost() {
-	const [request, setDialogState] = useState<SyncDialogRequest | null>(getCurrentRequest());
-
-	useEffect(() => {
-		setHostRequestSetter(setDialogState);
-		return () => {
-			if (isHostRequestSetter(setDialogState)) setHostRequestSetter(null);
-		};
-	}, []);
-
-	const open = Boolean(request);
-	const titleId = useMemo(() => `sync-dialog-title-${request?.kind || "none"}`, [request?.kind]);
-	const descriptionId = useMemo(
-		() => `sync-dialog-description-${request?.kind || "none"}`,
-		[request?.kind],
-	);
-
-	if (!request) return null;
-
-	const handleOpenChange = (nextOpen: boolean) => {
-		if (nextOpen) return;
-		resolveCurrentDialog(request?.kind === "confirm" ? false : null);
-	};
-
-	const handleOpenAutoFocus = (event: Event) => {
-		const legacyPrimary = document.querySelector<HTMLElement>(
-			'#syncDialog [data-sync-primary-action="true"]',
-		);
-		if (request?.kind === "input" && legacyPrimary) {
-			event.preventDefault();
-			legacyPrimary.focus();
-			return;
-		}
-		const preferredAction =
-			request?.kind === "confirm" ? (request.autoFocusAction ?? "confirm") : "confirm";
-		const selector =
-			preferredAction === "cancel"
-				? '#syncDialog [data-sync-dialog-action="cancel"]'
-				: '#syncDialog [data-sync-dialog-action="confirm"]';
-		const primary = document.querySelector<HTMLElement>(selector);
-		if (!primary) return;
-		event.preventDefault();
-		primary.focus();
-	};
-
-	const body = request ? (
-		<SyncDialogBody descriptionId={descriptionId} request={request} titleId={titleId} />
-	) : null;
-
-	return (
-		<RadixDialog
-			ariaDescribedby={descriptionId}
-			ariaLabelledby={titleId}
-			contentClassName="modal"
-			contentId="syncDialog"
-			onOpenAutoFocus={handleOpenAutoFocus}
-			onOpenChange={handleOpenChange}
-			open={open}
-			overlayClassName="modal-backdrop"
-			overlayId="syncDialogBackdrop"
-		>
-			{body}
-		</RadixDialog>
-	);
-}
-
-function SyncDialogBody({
-	descriptionId,
-	request,
-	titleId,
-}: {
-	descriptionId: string;
-	request: SyncDialogRequest;
-	titleId: string;
-}) {
-	const [inputValue, setInputValue] = useState(
-		request.kind === "input" ? request.initialValue || "" : "",
-	);
-	const [errorText, setErrorText] = useState<string | null>(null);
-	const supportsLabels = request.kind !== "duplicate-person";
-	const inputErrorId = "syncDialogInputError";
-
-	useEffect(() => {
-		if (request.kind === "input") {
-			setInputValue(request.initialValue || "");
-			setErrorText(null);
-		}
-	}, [request]);
-
-	const cancelLabel = supportsLabels ? request.cancelLabel || "Cancel" : "Cancel";
-	const confirmLabel = supportsLabels
-		? request.confirmLabel || (request.kind === "confirm" ? "Confirm" : "Save")
-		: "Confirm";
-
-	const submit = () => {
-		if (request.kind === "confirm") {
-			resolveCurrentDialog(true);
-			return;
-		}
-		if (request.kind === "duplicate-person") return;
-		const trimmed = inputValue.trim();
-		const validation = request.validate?.(trimmed) || null;
-		if (validation) {
-			setErrorText(validation);
-			return;
-		}
-		resolveCurrentDialog(trimmed);
-	};
-
-	return (
-		<div className="modal-card sync-dialog-card">
-			<div className="modal-header">
-				<h2 id={titleId}>{request.title}</h2>
-				<DialogCloseButton
-					ariaLabel={`Close ${request.title}`}
-					onClick={() => resolveCurrentDialog(request.kind === "confirm" ? false : null)}
-				/>
-			</div>
-			<div className="modal-body">
-				{request.kind !== "duplicate-person" ? (
-					<div className="small" id={descriptionId}>
-						{request.description}
-					</div>
-				) : null}
-				{request.kind === "duplicate-person" ? (
-					<div className="small" id={descriptionId}>
-						{request.summary}
-					</div>
-				) : null}
-				{request.kind === "input" ? (
-					<div className="field">
-						<TextInput
-							aria-describedby={errorText ? `${descriptionId} ${inputErrorId}` : descriptionId}
-							autoFocus
-							className={errorText ? "sync-dialog-input sync-field-error" : "sync-dialog-input"}
-							data-sync-primary-action="true"
-							onInput={(event) => {
-								setInputValue(event.currentTarget.value);
-								if (errorText) setErrorText(null);
-							}}
-							onKeyDown={(event) => {
-								if (event.key !== "Enter") return;
-								event.preventDefault();
-								submit();
-							}}
-							placeholder={request.placeholder}
-							type="text"
-							value={inputValue}
-						/>
-						{errorText ? (
-							<div className="sync-field-hint" id={inputErrorId}>
-								{errorText}
-							</div>
-						) : null}
-					</div>
-				) : null}
-				{request.kind === "duplicate-person" ? (
-					<DuplicatePersonDialogContent descriptionId={descriptionId} request={request} />
-				) : null}
-			</div>
-			<div className="modal-footer">
-				<div className="small" />
-				<div className="sync-dialog-actions">
-					<button
-						className="settings-button"
-						data-sync-dialog-action="cancel"
-						onClick={() => resolveCurrentDialog(request.kind === "confirm" ? false : null)}
-						type="button"
-					>
-						{cancelLabel}
-					</button>
-					{request.kind !== "duplicate-person" ? (
-						<button
-							autoFocus
-							className={`settings-button ${dialogToneClassName(request.kind === "confirm" ? request.tone : undefined)}`}
-							data-sync-dialog-action="confirm"
-							onClick={submit}
-							type="button"
-						>
-							{confirmLabel}
-						</button>
-					) : null}
-				</div>
-			</div>
-		</div>
-	);
-}
-
-function DuplicatePersonDialogContent({
-	descriptionId,
-	request,
-}: {
-	descriptionId: string;
-	request: DuplicatePersonDialogRequest;
-}) {
-	const [step, setStep] = useState<"choice" | "merge">("choice");
-	const defaultPrimary =
-		request.actors.find((actor) => actor.isLocal)?.actorId || request.actors[0]?.actorId || "";
-	const [primaryActorId, setPrimaryActorId] = useState(defaultPrimary);
-	const [secondaryActorId, setSecondaryActorId] = useState("");
-
-	useEffect(() => {
-		setStep("choice");
-		const nextPrimaryActorId =
-			request.actors.find((actor) => actor.isLocal)?.actorId || request.actors[0]?.actorId || "";
-		const nextSecondaryActorId =
-			request.actors.find((actor) => actor.actorId !== nextPrimaryActorId)?.actorId || "";
-		setPrimaryActorId(nextPrimaryActorId);
-		setSecondaryActorId(nextSecondaryActorId);
-	}, [request]);
-
-	const primary =
-		request.actors.find((actor) => actor.actorId === primaryActorId) || request.actors[0];
-	const mergeCandidates = request.actors.filter((actor) => actor.actorId !== primaryActorId);
-	const secondary =
-		mergeCandidates.find((actor) => actor.actorId === secondaryActorId) || mergeCandidates[0];
-	const primaryOptions = request.actors.map((actor) => ({
-		label: (
-			<>
-				{actor.label}
-				{actor.isLocal ? " (You)" : ""}
-			</>
-		),
-		value: actor.actorId,
-	}));
-	const mergeOptions = mergeCandidates.map((actor) => ({
-		label: `${actor.label}${actor.isLocal ? " (You)" : ""}`,
-		value: actor.actorId,
-	}));
-
-	return step === "choice" ? (
-		<div className="sync-dialog-stack">
-			<ul className="sync-dialog-choice-list">
-				<li>
-					<button
-						autoFocus
-						className="settings-button"
-						data-sync-primary-action="true"
-						onClick={() => setStep("merge")}
-						type="button"
-					>
-						These are both me
-					</button>
-				</li>
-				<li>
-					<button
-						className="settings-button"
-						onClick={() => resolveCurrentDialog({ action: "different-people" })}
-						type="button"
-					>
-						These are different people
-					</button>
-				</li>
-				<li>
-					<button
-						className="settings-button"
-						onClick={() => resolveCurrentDialog({ action: "cancel" })}
-						type="button"
-					>
-						Decide later
-					</button>
-				</li>
-			</ul>
-		</div>
-	) : (
-		<div className="sync-dialog-stack">
-			<div className="small" id={descriptionId}>
-				Choose which person should remain after combining these duplicates.
-			</div>
-			<RadixRadioGroup
-				ariaDescribedby={descriptionId}
-				ariaLabel="Person to keep after combining duplicates"
-				autoFocusValue={primaryActorId}
-				indicatorClassName="sync-dialog-radio-indicator"
-				itemClassName="sync-dialog-radio-option"
-				itemLabelClassName="sync-dialog-radio-label"
-				name="syncDuplicatePrimaryActor"
-				onValueChange={setPrimaryActorId}
-				options={primaryOptions}
-				rootClassName="sync-dialog-radio-list"
-				value={primaryActorId}
-			/>
-			<div className="field">
-				<label className="small" htmlFor="syncDuplicateSecondaryActor">
-					Person to combine into the selected record
-				</label>
-				<RadixSelect
-					ariaLabel="Person to combine into the selected record"
-					contentClassName="sync-radix-select-content sync-actor-select-content"
-					disabled={mergeOptions.length === 0}
-					id="syncDuplicateSecondaryActor"
-					itemClassName="sync-radix-select-item"
-					onValueChange={setSecondaryActorId}
-					options={mergeOptions}
-					placeholder="No merge target available"
-					triggerClassName="sync-radix-select-trigger sync-dialog-input"
-					value={secondary?.actorId || ""}
-					viewportClassName="sync-radix-select-viewport"
-				/>
-			</div>
-			<div className="sync-dialog-actions">
-				<button className="settings-button" onClick={() => setStep("choice")} type="button">
-					Back
-				</button>
-				<button
-					className="settings-button"
-					disabled={!primary?.actorId || !secondary?.actorId}
-					onClick={() => {
-						if (!primary?.actorId || !secondary?.actorId) return;
-						resolveCurrentDialog({
-							action: "merge",
-							primaryActorId: primary.actorId,
-							secondaryActorId: secondary.actorId,
-						});
-					}}
-					type="button"
-				>
-					Combine people
-				</button>
-			</div>
-		</div>
-	);
-}
 
 export function ensureSyncDialogHost() {
 	if (dialogMount?.isConnected) return;
